书评相关基本功能

+++ talk-server/src/main/java/com/cc/talkserver/user/service/BookUserService.java
@@ -55,7 +55,7 @@ public interface BookUserService extends IService<Book> {
      * @param id
      * @return
      */
-    PageResult<BookShowDTO> getPageByTag(Integer id,PageBookDTO pageBookDTO);
+    PageResult<BookShowDTO> getPageByTag(Long id,PageBookDTO pageBookDTO);
 
 
     /**
